feat: partially move signal to axsignal
