+++ Cargo.toml
@@ -37,6 +37,8 @@ arceos_posix_api = { git = "https://github.com/oscomp/arceos.git", features = ["
 axns = { git = "https://github.com/oscomp/arceos.git", features = ["thread-local"] }
 axptr = { git = "https://github.com/Starry-OS/axptr.git" }
 
+axsignal = { path = "/home/mivik/Projects/axsignal" }
+
 [patch.crates-io]
 syscalls = { git = "https://github.com/jasonwhite/syscalls.git", rev = "92624de"}
 page_table_multiarch = { git = "https://github.com/Mivik/page_table_multiarch.git", rev = "19ededd" }
